[alogirhtms] move contains_zero_vector here from float_cmp
