+++ src/pymor/tools/__init__.py
@@ -2,3 +2,6 @@
 # Copyright 2013-2020 pyMOR developers and contributors. All rights reserved.
 # License: BSD 2-Clause License (http://opensource.org/licenses/BSD-2-Clause)
 
+""" pymor.core.tools collects modules used throughout pyMOR
+that generally do not depend on any of pyMOR's abstraction objects.
+"""
